basic timetable

+++ src/main/java/com/hddev244/timetable/repository/BlockRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hddev244.timetable.entity.BlockEntity;
 
-public interface BlockRepository extends JpaRepository< BlockEntity, Integer> {
+public interface BlockRepository extends JpaRepository< BlockEntity, Long> {
 
 }
